# performance: avoid use of slow numpy.find_common_type