--- a/cupyx/scipy/sparse/linalg/solve.py
+++ b/cupyx/scipy/sparse/linalg/solve.py
@@ -43,7 +43,7 @@ def lsqr(A, b):
     if A.dtype == 'f' or A.dtype == 'd':
         dtype = A.dtype
     else:
-        dtype = numpy.find_common_type((A.dtype, 'f'), ())
+        dtype = numpy.promote_types(A.dtype, 'f')
 
     handle = device.get_cusolver_sp_handle()
     nnz = A.nnz