I have replaced `numpy.find_common_type` with use of `numpy.promote_types` where possible.

A couple basic benchmarks showing that relative timings (for NumPy 1.17.3):

e.g.
```
import numpy as np
%timeit np.find_common_type([np.int, 'f'], [])
6.4 µs ± 23.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

%timeit np.result_type(*(np.int, 'f'))
657 ns ± 1.19 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

%timeit np.promote_types(np.int, 'f')
148 ns ± 0.644 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

This is also true for long lists of types
```
import functools

import numpy as np

types = [np.int, 'f'] * 100

%timeit np.find_common_type(types, [])
84.8 µs ± 232 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit functools.reduce(np.promote_types, types)
12.2 µs ± 31.4 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

# with a large number of types it helps to call set first
%timeit functools.reduce(np.promote_types, set(types))
1.38 µs ± 2.32 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

```