@grlee77 2019-11-23
I forgot to do the same for cupyx. will add that here as well